pyproject.toml
[tool.poetry]
name = "que-vemos"
version = "0.1.0"
description = ""
authors = ["Alex Menor <alex4menor@gmail.com>"]
packages = [
  {include = "app"}
]

[tool.poetry.dependencies]
python = "^3.8"
uvicorn = "^0.12.3"
fastapi = "^0.62.0"
python-dotenv = "^0.15.0"
etcd3 = "^0.12.0"

[tool.poetry.dev-dependencies]
pylint = "^2.6.0"
pytest = "^6.1.1"
taskipy = "^1.4.0"
pytest-cov = "^2.10.1"
mutmut = "^2.1.0"
pytest-asyncio = "^0.14.0"
requests = "^2.25.0"

[tool.taskipy.tasks]
test = 'pytest'
cov = 'pytest --cov-report=xml --cov=app --cov-config=.coveragerc'
lint = 'pylint app'
mut = 'mutmut run --paths-to-mutate=app/ --paths-to-exclude=app/tests/ --tests-dir=app/tests/'
refresh-data = 'python -m app.data.watchables_extractor.watchables_extractor'
start = 'uvicorn app.app:app'
dev = 'uvicorn app.app:app --reload'

[build-system]
requires = ["poetry-core>=1.0.0a5"]
build-backend = "poetry.core.masonry.api"
